Refactor meals#show function to move and encapsulate logic into appropriate directories
