+++ lib/controllers/foodsController.js
@@ -61,7 +61,7 @@ const create = (request, response) => {
   }
   Food.findNewFood(food)
     .then(count => {
-      console.log(count)
+      // console.log(count)
       if (!(count[0]['count'] == "0")) {
         response.status(409).json({
           error: 'Duplicate entries are not permitted.'
@@ -95,7 +95,6 @@ const destroy = (request, response) => {
   }) 
 }
 
-  module.exports = {
-  index, show, update, destroy,
-  create,
+module.exports = {
+  index, show, update, create, destroy,
 }
