+++ lib/controllers/mealsController.js
@@ -1,4 +1,5 @@
 const Meal = require('../models/meal')
+const mealFood = require('../models/mealFood')
 
 const index = (request, response) => {
   Meal.allMeals()
@@ -7,6 +8,22 @@ const index = (request, response) => {
   })
   .catch((error) => {
     response.sendStatus(404).json({ error })
+  })
+};
+
+const show = (request, response) => {
+  const id = request.params.meal_id
+  mealFood.allMealFoods(id)
+  .then((foods) => {
+    if (foods.rows.length == 0) {
+      response.status(404).send({error: `Could not find meal with id ${id}`})
+    }
+    else {
+      response.status(200).json(foods.rows[0])
+    }
+  })
+  .catch((error) => {
+    response.status(500).json({ error })
   })
 };
 
@@ -43,5 +60,5 @@ const create = (request, response) => {
 };
 
 module.exports = {
-  index, create
+  show, index, create
 }
