+++ lib/models/food.js
@@ -23,6 +23,5 @@ const removeFood = (foodId) => database('foods')
 
 
 module.exports = {
-  allFoods, oneFood, editFood, removeFood,
-  findNewFood, newFood,
+  allFoods, oneFood, editFood, removeFood, findNewFood, newFood,
 }
