+++ lib/routes/api/v1/foods.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const foodsController = require('../../../controllers/foods_controller')
+const foodsController = require('../../../controllers/foodsController')
 
 router.get('/', foodsController.index);
 router.get('/:id', foodsController.show);
