+++ lib/routes/api/v1/meals.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const mealsController = require('../../../controllers/meals_controller')
+const mealsController = require('../../../controllers/mealsController')
 
 router.get('/', mealsController.index)
 router.post('/', mealsController.create)
+router.get('/:meal_id/foods', mealsController.show)
 
 module.exports = router
